Move DOM functions to a separate file.
Also, cleans up content_script.js.
